% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_polygon.r
\name{check_polygon}
\alias{check_polygon}
\title{Test for simple polygon}
\usage{
check_polygon(xy)
}
\arguments{
\item{xy}{2D matrix with x-y coordinates of vertices.}
}
\value{
If
}
\description{
It checks whether the sides of a general polygon do not intersect each other.
}
\details{
Calculations are...
}
\examples{

## A simple polygon.
x <- cumsum(runif(10))
y <- cumsum(runif(10))*x^4
plot(x,y,type="l",lty=1,lwd=2)
points(c(x[10],x[1]),c(y[10],y[1]),type="l",lty=2,lwd=2)
p <- check_polygon(cbind(x,y))

}
