% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments_intersection.r
\name{segments_intersection}
\alias{segments_intersection}
\title{Check whether two line segments cross each other.}
\usage{
segments_intersection(s1, s2, include_limits = T)
}
\arguments{
\item{s1}{2x2 matrix with x-y coordinates of vertices of first segment in first and second columns, respectively.}

\item{s2}{2x2 matrix with x-y coordinates of vertices of second segment in first and second columns, respectively.}

\item{include_limits}{Logical determines whether an intersection between the ends of the line segments is valid.}
}
\value{
Coordinates x-y of intersection point, if it exists, or NA if lines are parallel, or
FALSE if lines are not parallel but they do not intersect within the limits of both segments.
}
\description{
The line segments defined by the two input matrices are intersected.
}
\details{
It first checks whether lines are parallel, in which case they either have no common point or infinite.
Then it determines the intersection between both lines and checks whether that intersection takes place within
the limits of both segments.
}
\examples{

## An example of intersection.
s1 <- matrix(c(0,1,0,1),2,2)
s2 <- matrix(c(0,1,1,0),2,2)
plot(s1,type="l",lwd=2,xlab="x",ylab="y")
points(s2,type="l",lty=2,lwd=2)
xp <- segments_intersection(s1,s2)
points(xp[1],xp[2],cex=2)

## An example of segments that do not intersect.
plot(s1+3,type="l",lwd=2,xlab="x",ylab="y",xlim=c(0,4),ylim=c(0,4))
points(s2,type="l",lty=2,lwd=2)
segments_intersection(s1+3,s2)

# An example of parallel lines.
s1 <- matrix(c(1,2,1,2),2,2)
s2 <- matrix(c(0,3,1,4),2,2)
plot(s1,type="l",lwd=2,xlab="x",ylab="y",xlim=c(0,4),ylim=c(0,4))
points(s2,type="l",lty=2,lwd=2)
segments_intersection(s1,s2)

}
