% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_simu.R
\name{variogram_simu}
\alias{variogram_simu}
\title{Simulations of spatially-filtered variograms}
\usage{
variogram_simu(
  df,
  variable = variable,
  log_transform = TRUE,
  min_dist = min_dist,
  boundaries = boundaries,
  nsimu = 100,
  crs = NA,
  probs = c(0.025, 0.5, 0.975),
  verbose = TRUE
)
}
\arguments{
\item{df}{\code{data.frame} containing coordinates and variable values at those points.}

\item{variable}{single \code{character} indicating the name of the column in \code{data.frame} 'df' to calculate its variogram.}

\item{log_transform}{\code{logical} variable indicating whether 'variable' should be log-transformed.}

\item{min_dist}{\code{numeric} positive value specifying the minimum cutoff distance.}

\item{boundaries}{\code{numeric} vector with the limits of the distance intervals to calculate variogram.}

\item{nsimu}{\code{numeric} number of stochastic simulations (i.e. calls to 'distance_filter') to perform.}

\item{crs}{coordinate reference system; see help for \code{sf::st_as_sf}.}

\item{probs}{\code{numeric} vector of probabilities. See help for \code{quantile}.}

\item{verbose}{\code{logical} flag. If set to TRUE a progress bar will be shown on screen.}
}
\value{
A \code{data.frame} containing the average variogram at several distance intervals, as specified with
input parameter 'boundaries'. See example below.
}
\description{
\code{variogram_simu} spatially-filters (with \code{distance_filter} and then computes variograms
(with \code{variogram} from the 'gstat' package) for a number of simulations.
}
\details{
Function 'distance_filter' is used to calculate 'nsimu' variogram values inside distance intervals.
Next, mean and quantiles are computed within each of those intervals.
}
\examples{

# install.packages(c("sf", "gstat", "sp))
library(sp)
data(meuse)

# First we plot a variogram with all points.
cutoff <- 4000
boundaries <- seq(0, 4000, by = 200)
meu <- sf::st_as_sf(meuse, coords = c("x", "y"), crs = NA)
vmeuse <- gstat::variogram(log(zinc) ~ 1, data = meu, boundaries = boundaries, cutoff = cutoff)
plot(vmeuse, type="l", lwd=2,xlim=c(0, 4000), ylim=c(0, 1.5))

# We now simulate 1000 filtered meuse data sets. Minimum distance will be 500. Notice that the "confidence
# band" will vary depending on the width of intervals.
nsimu <- 1000
min_dist <- 500
interval_width <- 500
boundaries <- seq(min_dist, cutoff, by = interval_width)
r <- variogram_simu(meuse, variable = "zinc", log_transform = TRUE, min_dist = min_dist, boundaries = boundaries, nsimu = nsimu)
plot(r$mean_dist, r$variogram, type="l", lwd=2,xlim=c(0, 4000), ylim=c(0, 1.5), xlab = "Distance")
points(r$mean_dist, r$quantile_0.5, type = "l", lwd=2, lty = 2, col = "red")
points(r$mean_dist, r$quantile_0.025, type = "l", lwd=2, lty = 3, col = "red")
points(r$mean_dist, r$quantile_0.975, type = "l", lwd=2, lty = 3, col = "red")
}
