% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_seg_length.r
\name{calc_seg_length}
\alias{calc_seg_length}
\title{Total length of curve}
\usage{
calc_seg_length(s)
}
\arguments{
\item{s}{data.frame with at least two columns named "\code{x}" and "\code{y}" containing
the cartesian coordinates of segments' points. Segments must be consecutive.}
}
\value{
a number giving the total length.
}
\description{
It calculates the total length of a curve made up of a series of segments.
}
\details{
Total length is computed by adding the lengths of the individual segments.
}
\examples{
x <- runif(10)
y <- rnorm(10)
s <- data.frame(x=x,y=y)
calc_seg_length(s)

}
